Add a new not-really-a-Next-button needle for gnome-initial-setup
In g-i-s 3.37.91, the first screen has a 'Start Setup' button
rather than a 'Next' button. Easiest thing for us to do here is
just to add a new needle which has the 'next_button' tag even
though it's clearly not a 'Next' button, because then the code
still works :) So do that, but give the file a suggestive name
and explain the situation in a code comment.

Signed-off-by: Adam Williamson <awilliam@redhat.com>