Workaround RHBZ #1872511 by installing langpack
There's a complex bug in current Rawhide affecting the database
server test; it boils down to "deployment fails because LANG is
set to a locale for which the corresponding langpack is not
installed". As we know broadly what's going on there now, let's
work around it with a soft failure so we catch any later bugs in
the process.

Signed-off-by: Adam Williamson <awilliam@redhat.com>