Simplify anacondatest root_console to always use tty3
Rawhide KDE lives now have the desktop on tty2, and the installer
environment tty3 now has a shell (in Ye Olde Times it didn't, not
sure when that changed but it's the case at least back to F31).
So let's make our lives simple and just always use tty3 here.

Signed-off-by: Adam Williamson <awilliam@redhat.com>