Use PREVREL image not CURRREL for resize_lvm tests (#201)
These tests assume the desktop base disk image has an ext4-on-LVM
layout, but from F33 onwards it doesn't, it uses btrfs as that's
the new distro default.

We need a proper fix for this, but for now, just make the test
use the F32 disk image. This buys us six months at least.

Signed-off-by: Adam Williamson <awilliam@redhat.com>