desktop_notifications: wait longer at login screen
This matches the wait in boot_to_login_screen. The needle can
match before the UI is really done loading, and if we don't wait
long enough we wind up hitting enter before GDM is really ready
for us. This seems to be affecting the test on aarch64 quite
badly.

Signed-off-by: Adam Williamson <awilliam@redhat.com>