Handle KDE defaulting to offline updates in F34+
We need to hit 'restart' after applying updates, and we also
need the 'done' needle *not* to match the restart message, so
change that to match on the text (unfortunately). That also
means we have to add another variant of the needle for F32 as
the background of the text is a different color there.

Signed-off-by: Adam Williamson <awilliam@redhat.com>