Set several extra schema keys for update notification test
GNOME got even more clever-clever about only checking for and
notifying about updates after a certain amount of time, so we
need to fake it out even harder. See
https://bugzilla.redhat.com/show_bug.cgi?id=1930401 and
https://wiki.gnome.org/Design/Apps/Software/Updates#Tentative_Design
Note the test will still fail for now as there is an actual bug
that needs fixing, but once the fix is in, this should work.

Signed-off-by: Adam Williamson <awilliam@redhat.com>