Add an 'unwanted packages' test, run on Workstation for now (#218)
This adds a test that just fails if any one of a given list of
unwanted packages is installed. This was a request for the
Workstation edition from @catanzaro so I've just implemented it
for Workstation so far, but it's designed to be easily extended
to cover other subvariants too if we want.

Signed-off-by: Adam Williamson <awilliam@redhat.com>