Update leftbar generic needle, add rhbz#1959160 workaround
The background color of the generic left bar needle has changed
to match the new logo base color. The top bar's background color
has similarly changed, but this also caused us to notice a bug
in fedora-logos - that topbar image file seems to be basically
empty (just a transparent rectangle) so we see no 'image' in the
top bar, just solid electric blue. This needle matches that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>