Fix Cloud tests to run on correct image
For Cloud, we want to run these tests directly on the disk image
from the compose. But for other flavors, they are run on a disk
image produced by install_default_upload, so the test suites
specify HDD_1. This causes a problem as the value from the test
suite is used as the filename when downloading the image, but
that file name does not change between composes, so instead of
downloading the image to be tested for each compose, we just
wound up downloading it one time and then re-using that same
file every day.

Solving this is a bit tricky for reasons explained in the
fedora_openqa commit, but this is the best option I could think
of. The scheduler has been changed to schedule the downloaded
image as HDD_2_URL, not HDD_1_URL; so now in the templates we just
override the HDD_1 value for the Cloud flavors to "%HDD_2%",
meaning to take the value of HDD_2 (which will be parsed from
HDD_2_URL). We do not actually attach HDD_2 at all, it's only
used to be copied to HDD_1.

We also explicitly set DEPLOY_UPLOAD_TEST to "" for all flavors
(it was only set for one before), just for clarity.

Signed-off-by: Adam Williamson <awilliam@redhat.com>