Update Konsole runs needle, maximize Konsole in desktop_terminal
The needle match seems to have changed when bug #1957858 showed
up, but it's actually just a text rendering change in the window
title, it's not exactly caused by the tiny window. So not marking
as a workaround needle.

Maximizing the window makes the test work faster when we hit that
bug, as type_safely needs to be able to see the results of its
typing.

Signed-off-by: Adam Williamson <awilliam@redhat.com>