Manually add Japanese input source for install_asian_language
GNOME dropped the g-i-s new user mode in F34, so on a Japanese
install with user created in the installer, you don't get an
input source configured out of the box or on first boot. So
we'll just have to do it manually after booting, before we test
if it works.

Signed-off-by: Adam Williamson <awilliam@redhat.com>