Add default bug report and feature request issue templates (#44)
* Add default bug report and feature request issue templates

Using admin privs and settings->features->issue templates per [Github documentation](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) add default Issue templates to the repository.

* Rename openqa-test-suite-needle-update.md

Rename openqa-test-suite-needle-update.md to openqa_update.md as attempt to restore issue template picker.  This rename both shortens the name of the template and replaces `-` characters with `_`. Likely won't make a difference but it's worth a try.