Needle for server_realmd_join_kickstart
In module _boot_to_anaconda.pm line 93 there is a 60 min timeout
giving time for the bootloader, install and login sequence to complete.
This needle detects the end of that sequence and thereby allows the
server_realmd_join_kickstart test to complete which in turn allows the
server_role_deploy_domain_controller test also to complete.